+++ packages/components/modules/payments/web/ProductGallery/ProductDrawer/DrawerCarousel.tsx
