+++ packages/components/modules/payments/web/ProductGallery/ProductDrawer/DrawerProductDescription.tsx
